Update Gender property to use Byte instead of Int16
The `Gender` property in `clsPersonData.cs` within the
`DVLD_DataAccess` namespace was updated to use the `GetByte`
method instead of `GetInt16` when retrieving data from the
`reader`. This change modifies the data type of the `Gender`
property from `Int16` to `Byte`, aligning with the expected
data type and potentially optimizing memory usage.

+++ DVLD_DataAccess/clsPersonData.cs
@@ -38,7 +38,7 @@ namespace DVLD_DataAccess
                                 ThirdName = reader.IsDBNull(3) ? string.Empty : reader.GetString(3);
                                 LastName = reader.GetString(4);
                                 DateOfBirth = reader.GetDateTime(5);
-                                Gender = reader.GetInt16(6);
+                                Gender = reader.GetByte(6);
                                 Address = reader.GetString(7);
                                 Phone = reader.GetString(8);
                                 Email = reader.IsDBNull(9) ? string.Empty : reader.GetString(9);
